+++ packages/instant-apps-components/src/components/instant-apps-language-translator/support/interfaces.d.ts
@@ -6,9 +6,11 @@ interface AppSettings {
 interface LocaleSettingData {
   id: string;
   type: SettingType;
+  stringType: StringType;
   label: string;
   value: string;
   uiLocation?: any;
+  content?: LocaleSettingData[];
 }
 
 interface TranslatedLanguageLabels {
@@ -50,5 +52,10 @@ export interface LanguageTranslatorSearchResult extends LocaleSettingItem {
   fieldName: string;
 }
 
+export interface CalciteInputMap {
+  [id: string]: HTMLCalciteInputElement;
+}
+
 export type InputType = 'user' | 'translation';
 export type SettingType = 'string' | 'textEditor' | 'textarea';
+export type StringType = 'title' | 'subtitle' | 'text' | 'button' | 'string';